+++ backend/Watchdog.Core/Watchdog.Core.API/Controllers/UsersController.cs
@@ -17,13 +17,6 @@ namespace Watchdog.Core.API.Controllers
             _userService = userService;
         }
 
-        [HttpGet]
-        public async Task<ActionResult<ICollection<UserDto>>> GetAllAsync()
-        {
-            var users = await _userService.GetAllUsersAsync();
-            return Ok(users);
-        }
-
         [HttpGet("{id:int}")]
         public async Task<ActionResult<UserDto>> GetUserByIdAsync(int id)
         {
@@ -42,6 +35,18 @@ namespace Watchdog.Core.API.Controllers
         public async Task<ActionResult<UserDto>> CreateUserAsync(NewUserDto userDto)
         {
             var user = await _userService.CreateUserAsync(userDto);
+            return Ok(user);
+        }
+
+        [HttpPut("{userId}")]
+        public async Task<ActionResult<UserDto>> UpdateAsync(int userId, UpdateUserDto updateUserDto)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
+            var user = await _userService.UpdateUserAsync(userId, updateUserDto);
             return Ok(user);
         }
     }
