+++ backend/Watchdog.Core/Watchdog.Core.BLL/Services/Abstract/IUserService.cs
@@ -5,11 +5,7 @@ using Watchdog.Core.Common.DTO.User;
 namespace Watchdog.Core.BLL.Services.Abstract
 {
     public interface IUserService
-    {
-        Task<ICollection<UserDto>> GetAllUsersAsync();
-        
-        Task<UserDto> GetUserAsync(int userId);
-        
+    {           
         Task<UserDto> UpdateUserAsync(int userId, UpdateUserDto updateUserDto);
 
         Task<UserDto> GetUserByIdAsync(int id);
