+++ backend/Watchdog.Core/Watchdog.Core.Common/DTO/User/UpdateUserDto.cs
@@ -4,7 +4,6 @@ namespace Watchdog.Core.Common.DTO.User
 {
     public class UpdateUserDto
     {
-        public int Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
