+++ backend/Watchdog.Core/Watchdog.Core.Common/DTO/User/UserDto.cs
@@ -17,7 +17,7 @@ namespace Watchdog.Core.Common.DTO.User
         
         public string PasswordHash { get; set; }
         
-        public DateTime? RegisteredAt { get; set; }
+        public DateTime RegisteredAt { get; set; }
         
         public string AvatarUrl { get; set; }
     }
