+++ frontend/src/app/app-routing.module.ts
@@ -10,7 +10,7 @@ const routes: Routes = [
     },
     {
         path: 'home',
-        //canActivate: [AuthGuard],
+        canActivate: [AuthGuard],
         loadChildren: () => import('./modules/home/home.module')
             .then(m => m.HomeModule)
     },
