+++ frontend/src/app/core/services/user.service.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-import { UserUpdateDto } from '@core/models/userUpdate';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { User } from '@shared/models/user/user';
@@ -11,7 +10,7 @@ import { ShareDataService } from './share-data.service';
     providedIn: 'root'
 })
 export class UserService {
-    private apiPrefix = '/user';
+    private apiPrefix = '/users';
     private apiPrefixRegister = 'users';
 
     constructor(
@@ -24,7 +23,7 @@ export class UserService {
             .pipe(map(response => response.body));
     }
 
-    public updateUsersById(id: number, user: UserUpdateDto): Observable<UserUpdateDto> {
+    public updateUsersById(id: number, user: User): Observable<User> {
         return this.httpService.putFullRequest<User>(`${this.apiPrefix}/${user.id}`, user)
             .pipe(map(response => {
                 this.dataService.changeMessage(response.body);
