--- frontend/src/app/shared/models/user/userUpdate.ts
@@ -1,6 +1,0 @@
-export interface UserUpdateDto {
-    id: number;
-    email: string;
-    firstName: string;
-    lastName: string;
-}
